requirements.txt
flask
paho-mqtt
flask-socketio
requests